move new KDTree in kdtree_utils.py
